Follow convention when calling module methods
We have the convention of calling module methods using dot notation
instead of double colon notation.  Without this patch the two new
functions are "special" in that they don't follow this convention, which
is odd.

This patch fixes the problem by calling the
Puppet::Parser::Functions.newfunction method using the dot convention.